+++ packages/base/src/components/index.ts
@@ -1,4 +1,6 @@
+// SHOPLFLOW_DOCS_IGNORE
 export * from './Stack';
 export * from './Text';
 export * from './Modal';
 export * from './BackDrop';
+export * from './Chips';