+++ packages/base/src/providers/index.ts
@@ -1,3 +1,4 @@
+// SHOPLFLOW_DOCS_IGNORE
 export { default as ShoplflowProvider } from './ShoplflowProvider';
 export type { ShoplflowProviderProps } from './ShoplflowProvider';
 export { default as ModalProvider } from '../portal/ModalPortal';